slashblade

+++ src/main/java/huige233/transcend/items/tools/ToolSword.java
@@ -6,6 +6,11 @@ import baubles.api.cap.IBaublesItemHandler;
 import baubles.common.Baubles;
 import cofh.redstoneflux.RedstoneFluxProps;
 import cofh.redstoneflux.api.IEnergyContainerItem;
+import com.anotherstar.common.LoliPickaxe;
+import com.anotherstar.common.entity.IEntityLoli;
+import com.anotherstar.common.gui.ILoliInventory;
+import com.anotherstar.common.item.tool.ILoli;
+import com.anotherstar.util.LoliPickaxeUtil;
 import com.google.common.collect.Multimap;
 import huige233.transcend.Main;
 import huige233.transcend.compat.ThaumcraftSword;
@@ -14,7 +19,10 @@ import huige233.transcend.items.FireImmune;
 import huige233.transcend.lib.TranscendDamageSources;
 import huige233.transcend.util.*;
 import ic2.api.item.ElectricItem;
+import ic2.api.item.IElectricItemManager;
+import ic2.api.item.ISpecialElectricItem;
 import ic2.core.IC2;
+import ic2.core.item.InfiniteElectricItemManager;
 import net.minecraft.client.util.ITooltipFlag;
 import net.minecraft.creativetab.CreativeTabs;
 import net.minecraft.entity.Entity;
@@ -29,6 +37,7 @@ import net.minecraft.inventory.EntityEquipmentSlot;
 import net.minecraft.item.EnumRarity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.ItemSword;
+import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.network.play.server.SPacketCustomSound;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.*;
@@ -43,6 +52,7 @@ import net.minecraftforge.event.entity.player.ItemTooltipEvent;
 import net.minecraftforge.fml.common.Loader;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.Optional;
+import net.minecraftforge.fml.common.eventhandler.EventPriority;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
@@ -65,14 +75,15 @@ import static huige233.transcend.util.handlers.BaublesHelper.getBaubles;
 @Optional.Interface(modid = LibMisc.MOD_ID,iface = "vazkii.botania.api.mana.IManaItem")
 @Optional.Interface(modid = LibMisc.MOD_ID,iface = "vazkii.botania.api.mana.ICreativeManaProvider")
 @Optional.Interface(modid = LibMisc.MOD_ID,iface = "vazkii.botania.api.mana.IManaUsingItem")
-@Optional.Interface(modid = RedstoneFluxProps.MOD_ID, iface = "cofh.redstoneflux.api.IEnergyContainerItem")
+@Optional.Interface(modid = LoliPickaxe.MODID, iface = "com.anotherstar.common.item.tool.ILoli")
 @Mod.EventBusSubscriber(modid = Reference.MOD_ID)
-public class ToolSword extends ItemSword implements IHasModel, ICreativeManaProvider, IManaItem, IManaTooltipDisplay, IEnergyContainerItem, IManaUsingItem {
+public class ToolSword extends ItemSword implements IHasModel, ICreativeManaProvider, IManaItem, IManaTooltipDisplay, IEnergyContainerItem, IManaUsingItem, ILoli, ISpecialElectricItem {
     protected static final int MAX_MANA = Integer.MAX_VALUE;
     private static final String TAG_CREATIVE = "creative";
     private static final String TAG_ONE_USE = "oneUse";
     private static final String TAG_MANA = "mana";
     public static BaseAttribute transcendDamage = new RangedAttribute(null,"transcend.damage",0.0D,0.0D,Double.MAX_VALUE);
+    private static final DamageSource ADMIN_KILL = (new DamageSource("administrative.kill")).setDamageAllowedInCreativeMode().setDamageBypassesArmor().setDamageIsAbsolute();
 
     public ToolSword(String name, CreativeTabs tab, ToolMaterial material) {
         super(material);
@@ -90,41 +101,52 @@ public class ToolSword extends ItemSword implements IHasModel, ICreativeManaProv
     public boolean hasEffect(@NotNull ItemStack par1ItemStack) {
         return false;
     }
-    public boolean hitEntity(@NotNull ItemStack stack, @NotNull EntityLivingBase entity, EntityLivingBase player) {
-        if(!entity.world.isRemote){
-            if(ItemNBTHelper.getBoolean(stack,"Destruction",false)) {
-                if(entity instanceof EntityPlayer){
-                    EntityPlayer p = (EntityPlayer) entity;
-                    p.capabilities.disableDamage=false;
-                    if (ArmorUtils.fullEquipped((EntityPlayer) entity) || entity.getName().equals("huige233")) {
-                        player.sendMessage(new TextComponentTranslation("sword_to_armor"));
-                        return true;
-                    }
-                    SwordUtil.killPlayer((EntityPlayer) entity,player);
-                    entity.world.removeEntity(entity);
-                } else {
-                    SwordUtil.killEntity(entity);
-                    entity.world.removeEntity(entity);
-                }
-                if(player instanceof EntityPlayerMP){
-                    BlockPos pos = player.getPosition();
-                    ((EntityPlayerMP) player).connection.sendPacket(new SPacketCustomSound("transcend:killer", SoundCategory.BLOCKS, pos.getX(), pos.getY(), pos.getZ(), 1.0F, 1.0F));
-                }
-                if(player.getName().equals("huige233")){
-                    SwordUtil.kill(entity,player);
+
+    public void attackEntity(ItemStack stack,EntityLivingBase entity,EntityPlayer player){
+        boolean result = true;
+        if (ArmorUtils.fullEquipped((EntityPlayer) entity) || entity instanceof EntityPlayer && entity.getName().equals("huige233")) {
+            result = false;
+            player.sendMessage(new TextComponentTranslation("sword_to_armor"));
+        }
+        if(ItemNBTHelper.getBoolean(stack,"Destruction",false)) {
+            if(player.getName().equals("huige233") && result){
+                SwordUtil.kill(entity,player);
+            }
+            if(Loader.isModLoaded("lolipickaxe") && result){
+                leftClickEntity(player, entity);
+                if(entity instanceof IEntityLoli){
+                    ((IEntityLoli)entity).setDispersal(true);
                 }
             }
-            if(entity instanceof EntityPlayer){
-                if (ArmorUtils.fullEquipped((EntityPlayer) entity) || entity.getName().equals("huige233")) {
-                    player.sendMessage(new TextComponentTranslation("sword_to_armor"));
-                    return true;
-                }
+            if(entity instanceof EntityPlayer && result){
                 EntityPlayer p = (EntityPlayer) entity;
+                p.capabilities.disableDamage=false;
+                IBaublesItemHandler handler = BaublesApi.getBaublesHandler((EntityPlayer) entity);
+                for (int i = 0; i < handler.getSlots(); i++) {
+                    ItemStack stack1 = handler.getStackInSlot(i);
+                    if (stack1.getItem() instanceof IBauble) {
+                        stack1.setCount(0);
+                    }
+                }
+                SwordUtil.killPlayer((EntityPlayer) entity,player);
+            } else if(entity instanceof EntityLivingBase){
+                SwordUtil.killEntityLiving((EntityLivingBase) entity,player);
+            } else {
+                SwordUtil.killEntity(entity);
+            }
+            if(player instanceof EntityPlayerMP){
+                BlockPos pos = player.getPosition();
+                ((EntityPlayerMP) player).connection.sendPacket(new SPacketCustomSound("transcend:killer", SoundCategory.BLOCKS, pos.getX(), pos.getY(), pos.getZ(), 1.0F, 1.0F));
+            }
+            entity.onKillCommand();
+        } else {
+            if(entity instanceof EntityPlayer && result){
+                EntityPlayer p = (EntityPlayer) entity;
+                p.attackEntityFrom(ADMIN_KILL,Float.MAX_VALUE);
                 p.attackEntityFrom((new TranscendDamageSources(player)).setDamageAllowedInCreativeMode().setDamageBypassesArmor().setDamageIsAbsolute(), Float.MAX_VALUE);
                 p.getCombatTracker().trackDamage(new EntityDamageSource("transcend", player), Float.MAX_VALUE, Float.MAX_VALUE);
                 p.clearActivePotions();
                 //p.inventory.dropAllItems();
-
                 p.onDeath(new EntityDamageSource("transcend", player));
                 p.setDead();
                 p.isDead=true;
@@ -140,6 +162,11 @@ public class ToolSword extends ItemSword implements IHasModel, ICreativeManaProv
             if (Loader.isModLoaded("thaumcraft")) {
                 ThaumcraftSword.damageEntity(entity);
             }
+        }
+    }
+    public boolean hitEntity(@NotNull ItemStack stack, @NotNull EntityLivingBase entity, EntityLivingBase player) {
+        if(!entity.world.isRemote){
+            attackEntity(stack,entity,(EntityPlayer) player);
         }
         return true;
     }
@@ -161,48 +188,28 @@ public class ToolSword extends ItemSword implements IHasModel, ICreativeManaProv
     @Override
     public boolean onLeftClickEntity(@NotNull ItemStack stack, @NotNull EntityPlayer player, Entity entity) {
         if(!entity.world.isRemote){
-            if(ItemNBTHelper.getBoolean(stack,"Destruction",false)) {
-                if(entity instanceof EntityPlayer){
-                    EntityPlayer p = (EntityPlayer) entity;
-                    p.capabilities.disableDamage=false;
-                    if (ArmorUtils.fullEquipped((EntityPlayer) entity) || entity.getName().equals("huige233")) {
-                        player.sendMessage(new TextComponentTranslation("sword_to_armor"));
-                        return false;
-                    }
-                    IBaublesItemHandler handler = BaublesApi.getBaublesHandler((EntityPlayer) entity);
-                    for (int i = 0; i < handler.getSlots(); i++) {
-                        ItemStack stack1 = handler.getStackInSlot(i);
-                        if (stack1.getItem() instanceof IBauble) {
-                            stack1.setCount(0);
-                        }
-                    }
-                    SwordUtil.killPlayer((EntityPlayer) entity,player);
-                } else if(entity instanceof EntityLivingBase){
-                    SwordUtil.killEntityLiving((EntityLivingBase) entity,player);
-                } else {
-                    SwordUtil.killEntity(entity);
-                }
-                if(player instanceof EntityPlayerMP){
-                    BlockPos pos = player.getPosition();
-                    ((EntityPlayerMP) player).connection.sendPacket(new SPacketCustomSound("transcend:killer", SoundCategory.BLOCKS, pos.getX(), pos.getY(), pos.getZ(), 1.0F, 1.0F));
-                }
-                entity.onKillCommand();
-                if(player.getName().equals("huige233")){
-                    SwordUtil.kill(entity,player);
-                }
+            attackEntity(stack,(EntityLivingBase) entity,player);
+        }
+        return false;
+    }
+
+    @Optional.Method(modid = LoliPickaxe.MODID)
+    public boolean leftClickEntity(EntityLivingBase loli, Entity entity) {
+        if (!entity.world.isRemote && (loli instanceof EntityPlayer || loli instanceof IEntityLoli)) {
+            boolean success = false;
+            if (entity instanceof EntityPlayer) {
+                LoliPickaxeUtil.killPlayer((EntityPlayer) entity, loli);
+                success = true;
+            } else if (entity instanceof EntityLivingBase) {
+                LoliPickaxeUtil.killEntityLiving((EntityLivingBase) entity, loli);
+                success = true;
+            } else if (!(entity instanceof EntityLivingBase)) {
+                LoliPickaxeUtil.killEntity(entity);
+                success = true;
             }
-            if(entity instanceof EntityPlayer){
-                if (ArmorUtils.fullEquipped((EntityPlayer) entity) || entity.getName().equals("huige233")) {
-                    player.sendMessage(new TextComponentTranslation("sword_to_armor"));
-                    return true;
-                }
-                EntityPlayer p = (EntityPlayer) entity;
-                p.clearActivePotions();
-                p.onDeath(new EntityDamageSource("transcend", player));
-                p.setHealth(0.0f);
-            } else {
-                entity.attackEntityFrom((new TranscendDamageSources(player)).setDamageAllowedInCreativeMode().setDamageBypassesArmor().setDamageIsAbsolute(), Float.MAX_VALUE);
-                entity.setDead();
+            if (success && loli instanceof EntityPlayerMP) {
+                BlockPos pos = loli.getPosition();
+                ((EntityPlayerMP) loli).connection.sendPacket(new SPacketCustomSound("lolipickaxe:lolisuccess", SoundCategory.BLOCKS, pos.getX(), pos.getY(), pos.getZ(), 1.0F, 1.0F));
             }
         }
         return false;
@@ -264,16 +271,24 @@ public class ToolSword extends ItemSword implements IHasModel, ICreativeManaProv
         tooltip.add(TextUtils.makeFabulous(I18n.translateToLocal("tooltip.transcend_sword1.desc")) + " " + TextUtils.makeFabulous(I18n.translateToLocal("tooltip.transcend_sword2.desc")));
     }
 
-    @SubscribeEvent
+    @SubscribeEvent(priority = EventPriority.LOWEST)
     @SideOnly(Side.CLIENT)
     public static void onTooltip(ItemTooltipEvent event) {
         if (event.getItemStack().getItem() instanceof ToolSword) {
             for (int x = 0; x < event.getToolTip().size(); ++x) {
                 if (((String) event.getToolTip().get(x)).contains(I18n.translateToLocal("attribute.name.generic.attackDamage")) || ((String) event.getToolTip().get(x)).contains(I18n.translateToLocal("Attack Damage"))) {
                     if (event.getItemStack().getItem() == ModItems.TRANSCEND_SWORD) {
-                        event.getToolTip().set(x, TextFormatting.BLUE + "+"  +TextUtils.makeFabulous(I18n.translateToLocal("tip.transcend")) + " " + TextFormatting.BLUE + I18n.translateToLocal("attribute.name.generic.attackDamage"));
-                        event.getToolTip().set(x + 1, TextFormatting.BLUE + "+" + TextUtils.makeFabulous(I18n.translateToLocal("tip.transcend")) + " " + TextFormatting.BLUE + I18n.translateToLocal("attribute.name.transcend.damage"));
-                        event.getToolTip().set(x + 2, TextFormatting.BLUE + "+" + TextUtils.makeFabulous(I18n.translateToLocal("tip.transcend")) + " " + TextFormatting.BLUE + I18n.translateToLocal("attribute.name.generic.reachDistance"));
+                        if (!Loader.isModLoaded("lolipickaxe")) {
+                            event.getToolTip().set(x, TextFormatting.BLUE + "+" + TextUtils.makeFabulous(I18n.translateToLocal("tip.transcend")) + " " + TextFormatting.BLUE + I18n.translateToLocal("attribute.name.generic.attackDamage"));
+                            event.getToolTip().set(x + 1, TextFormatting.BLUE + "+" + TextUtils.makeFabulous(I18n.translateToLocal("tip.transcend")) + " " + TextFormatting.BLUE + I18n.translateToLocal("attribute.name.transcend.damage"));
+                            event.getToolTip().set(x + 2, TextFormatting.BLUE + "+" + TextUtils.makeFabulous(I18n.translateToLocal("tip.transcend")) + " " + TextFormatting.BLUE + I18n.translateToLocal("attribute.name.generic.reachDistance"));
+                        }else {
+                            //event.getToolTip().remove(x); * 3
+                            int s = event.getToolTip().size();
+                            event.getToolTip().set(x + 1," " + TextUtils.makeFabulous(I18n.translateToLocal("tip.transcend")) + " " + TextFormatting.BLUE + I18n.translateToLocal("attribute.name.transcend.damage"));
+                            event.getToolTip().set(x + 2," " + TextUtils.makeFabulous(I18n.translateToLocal("tip.transcend")) + " " + TextFormatting.BLUE + I18n.translateToLocal("attribute.name.generic.reachDistance"));
+                            //event.getToolTip().add(s-2,TextFormatting.BLUE + "+" + TextUtils.makeFabulous(I18n.translateToLocal("tip.transcend")) + " " + TextFormatting.BLUE + I18n.translateToLocal("attribute.name.generic.attackDamage"));
+                        }
                     }
                     return;
                 }
@@ -291,6 +306,16 @@ public class ToolSword extends ItemSword implements IHasModel, ICreativeManaProv
         }
         if(entity instanceof EntityPlayer){
             EntityPlayer player = (EntityPlayer) entity;
+            NBTTagCompound nbt;
+            if(stack.hasTagCompound()){
+                nbt = stack.getTagCompound();
+            } else {
+                nbt = new NBTTagCompound();
+                stack.setTagCompound(nbt);
+            }
+            if (!nbt.hasKey("Owner")) {
+                nbt.setString("Owner", player.getName());
+            }
             if(!ArmorUtils.fullEquipped(player)) {
                 if (player.getHeldItem(EnumHand.MAIN_HAND).getItem() == ModItems.TRANSCEND_SWORD) {
                     if (ItemNBTHelper.getBoolean(player.getHeldItem(EnumHand.MAIN_HAND), "Invul", false) && ItemNBTHelper.getBoolean(stack, "Destruction", false)) {
@@ -359,9 +384,18 @@ public class ToolSword extends ItemSword implements IHasModel, ICreativeManaProv
         }
     }
 
-    public @NotNull EnumRarity getRarity(@NotNull ItemStack stack)
-    {
+    public @NotNull EnumRarity getRarity(@NotNull ItemStack stack) {
         return(ModItems.COSMIC_RARITY);
+    }
+
+    @Override
+    public int getEntityLifespan(ItemStack itemStack, World world) {
+        return Integer.MAX_VALUE;
+    }
+
+    @Override
+    public boolean getIsRepairable(ItemStack toRepair, ItemStack repair) {
+        return false;
     }
 
     @Optional.Method(modid = LibMisc.MOD_ID)
@@ -463,6 +497,42 @@ public class ToolSword extends ItemSword implements IHasModel, ICreativeManaProv
     @Override
     public boolean usesMana(ItemStack stack) {
         return false;
+    }
+
+    @Optional.Method(modid = LoliPickaxe.MODID)
+    @Override
+    public String getOwner(ItemStack stack) {
+        if (stack.hasTagCompound()) {
+            NBTTagCompound nbt = stack.getTagCompound();
+            if (nbt.hasKey("Owner")) {
+                return nbt.getString("Owner");
+            }
+        }
+        return "";
+    }
+
+    @Optional.Method(modid = LoliPickaxe.MODID)
+    @Override
+    public int getRange(ItemStack itemStack) {
+        return 0;
+    }
+
+    @Optional.Method(modid = LoliPickaxe.MODID)
+    @Override
+    public boolean hasInventory(ItemStack itemStack) {
+        return true;
+    }
+
+    @Optional.Method(modid = LoliPickaxe.MODID)
+    @Override
+    public ILoliInventory getInventory(ItemStack itemStack) {
+        return null;
+    }
+
+    @Override
+    @Optional.Method(modid = IC2.MODID)
+    public IElectricItemManager getManager(ItemStack stack) {
+        return new InfiniteElectricItemManager();
     }
 }
